Remove unused file
